MCP-StreamableHttp/pyproject.toml
[project]
name = "mcp-streamablehttp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "mcp>=1.9.4",
    "tavily-python>=0.7.8",
]
